+++ src/commonMain/kotlin/util/control/NonFatal.kt
@@ -1,0 +1,39 @@
+/*
+ *
+ * This Kotlin code is based on Scala, licensed under the Apache License 2.0.
+ *
+ * ===== SCALA LICENSE =====
+ *
+ * Scala (https://www.scala-lang.org)
+ *
+ * Copyright EPFL and Lightbend, Inc.
+ *
+ * Licensed under Apache License 2.0
+ * (http://www.apache.org/licenses/LICENSE-2.0).
+ *
+ * See the NOTICE file distributed with this work for
+ * additional information regarding copyright ownership.
+ *
+ */
+
+package it.czerwinski.kotlin.util.control
+
+/**
+ * Matcher for non-fatal Throwables.
+ *
+ * _Note: [AssertionError] is not matched by `NonFatal`, as it's intended to be thrown iff
+ * the assumptions about the code are false._
+ *
+ * _Note: [NotImplementedError] is not matched by `NonFatal`, as it's intended to indicate
+ * a method that remains to be implemented._
+ *
+ * @since 1.4.30
+ */
+expect object NonFatal {
+
+    /**
+     * Returns `true` if the provided `Throwable` is to be considered non-fatal,
+     * or `false` if it is to be considered fatal.
+     */
+    operator fun invoke(exception: Throwable): Boolean
+}