+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Dependencies:
   - Upgrade `io.gitlab.arturbosch.detekt` to `1.15.0`
   - Upgrade `org.jetbrains.dokka` to `1.4.20`
+  - Upgrade `kotlin-multiplatform` to `1.4.21-2`
 
 ## [v1.4.21]
 ### Changed